refactor: separando clients (acesso a dados externos) de repositories (acesso a banco de dados)
