+++ backend/app/clients/__init__.py
@@ -1,0 +1,2 @@
+# Clients para comunicação com APIs externas
+
