+++ backend/tests/clients/__init__.py
@@ -1,0 +1,2 @@
+# Testes para clients de APIs externas
+
