+++ backend/tests/clients/myclimate_client/test_bulk_calculate_carbon_emission.py
@@ -3,7 +3,7 @@ import random
 
 import pytest
 from app.exceptions import MyclimateError
-from app.repositories.myclimate_client import (
+from app.clients.myclimate_client import (
     MAXIMUM_ACCEPTED_DISTANCE,
     MINIMUM_ACCEPTED_DISTANCE,
     bulk_calculate_carbon_emission,
