+++ backend/tests/clients/sptrans_client/test_get_lines.py
@@ -1,4 +1,4 @@
-from app.repositories.sptrans_client import get_lines
+from app.clients.sptrans_client import get_lines
 
 from tests.factories.schemas import SPTransLineFactory
 from tests.helpers import SPTransHelper
